requirements.txt
bcrypt==4.3.0
bleach==6.2.0
cffi~=1.17.1
configparser==7.2.0
flask==3.1.1
Flask-Caching==2.3.1
flask-socketio==5.5.1
gunicorn==23.0.0
jinja2==3.1.6
markdown==3.8
mysql-connector-python==9.3.0
#mysql==0.0.3
packaging==24.2
pillow==11.2.1
portalocker==3.1.1
pymysql==1.1.1
qrcode~=8.0
requests~=2.32.3
waitress~=3.0.2
Werkzeug~=3.1.3
user-agents==2.2.0
PyJWT==2.10.1
opencv-python~=4.11.0.86
python-dotenv==1.1.0
BeautifulSoup4==4.13.4
python-magic~=0.4.27
pytz~=2025.2
flask-siwadoc==0.2.2
libmagic
#windows python-magic-bin
prometheus_client~=0.22.1
flask-sqlalchemy==3.1.1
humanize==4.12.3
Flask-Bcrypt==1.0.1
mysqlclient==2.2.7
SQLAlchemy==2.0.42
psutil==7.0.0